requirements.txt
fastapi
uvicorn
httpx
jinja2
python-jose[cryptography]
cachetools
python-dotenv